backend_clean/requirements.txt
# Core FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.3.0
postgrest>=0.13.0

# Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4

# AI/ML
openai>=1.3.0
chromadb>=0.4.18

# File processing
PyPDF2>=3.0.1
python-docx>=1.1.0
Pillow>=10.1.0
pytesseract>=0.3.10

# HTTP requests (compatible with supabase)
httpx>=0.24.0,<0.25.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.1.0
